requirements.txt
discord.py
notion-client
apscheduler
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib